+++ example/App.js
@@ -36,7 +36,7 @@ import TealiumCrashReporter from 'tealium-react-native-crash-reporter';
 import TealiumAttribution from 'tealium-react-native-attribution';
 import AttributionConfig from 'tealium-react-native-attribution/common';
 import TealiumMomentsApi from 'tealium-react-native-moments-api';
-import { TealiumMomentsApiConfig, MomentsApiRegion} from 'tealium-react-native-moments-api/common';
+import { MomentsApiConfig, MomentsApiRegion} from 'tealium-react-native-moments-api/common';
 
 export default class App extends Component<{}> {
 
@@ -65,8 +65,8 @@ export default class App extends Component<{}> {
             iosSkAdConversionKeys: {"event": "conversion_value"}
         }
 
-        let momentsApiConfig: TealiumMomentsApiConfig = {
-            region: MomentsApiRegion.UsEast
+        let momentsApiConfig: MomentsApiConfig = {
+            momentsApiRegion: MomentsApiRegion.UsEast
         }
 
         // TealiumAdobeVisitor.configure(adobeVisitorConfig);