+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 [Full documentation](https://docs.tealium.com/platforms/react-native/install/)
 
-- 2.4.3
-  - Android fix: update VisitorProfile date types to align with iOS
+- 2.5.0
+  - Android fix: VisitorProfile dates now displayed as numerical timestamps instead of strings
   - Added `VisitorProfile` type definition
 - 2.4.2
   - Android fix: added check during initialization to ensure only a single instance of Tealium is running. Pre-existing instances are shutdown.